Move getEnvVar into src/kit
