+++ src/git/instance.ts
@@ -1,5 +1,5 @@
 import { Octokit } from "octokit";
-import { getEnvVar } from "@/env";
+import { getEnvVar } from "@/kit/env";
 
 export function getGit(): Octokit {
   return new Octokit({ auth: getEnvVar("MEMOS_GITHUB_TOKEN") });
