+++ src/vercel/domain.ts
@@ -1,5 +1,5 @@
 import { VercelError } from "@vercel/sdk/models/vercelerror.js";
-import { getEnvVar } from "@/env";
+import { getEnvVar } from "@/kit/env";
 import { getVercel } from "./instance";
 
 export async function addVercelDomain(params: { name: string }): Promise<void> {
@@ -19,4 +19,24 @@ export async function addVercelDomain(params: { name: string }): Promise<void> {
     idOrName,
     requestBody: { name },
   });
+}
+
+export type VercelDomainDns = {
+  cname: string | null;
+  ipv4: string | null;
+};
+
+export async function getVercelDomainDns(params: {
+  name: string;
+}): Promise<VercelDomainDns> {
+  const { name } = params;
+  const vercel = getVercel();
+  const projectIdOrName = getEnvVar("MEMOS_VERCEL_PROJECT_ID");
+  const config = await vercel.domains.getDomainConfig({
+    domain: name,
+    projectIdOrName,
+  });
+  const cname = config.recommendedCNAME.at(0)?.value ?? null;
+  const ipv4 = config.recommendedIPv4.at(0)?.value.at(0) ?? null;
+  return { cname, ipv4 };
 }
