+++ src/vercel/instance.ts
@@ -1,5 +1,5 @@
 import { Vercel } from "@vercel/sdk";
-import { getEnvVar } from "@/env";
+import { getEnvVar } from "@/kit/env";
 
 export function getVercel(): Vercel {
   return new Vercel({ bearerToken: getEnvVar("MEMOS_VERCEL_TOKEN") });
